api/pyproject.toml
[project]
name = "api"
version = "0.0.1"
description = "Api REST Demo"
requires-python = ">=3.12"
readme = "README.md"
license = "MIT"
authors = [{ name = "Luiz Otávio Miranda" }]

dependencies = [
  "fastapi>=0.116.1,<1.0.0",
  "opentelemetry-api>=1.38.0,<2.0.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.38.0,<2.0.0",
  "opentelemetry-instrumentation-fastapi>=0.59b0,<1.0.0",
  "opentelemetry-sdk>=1.38.0,<2.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.scripts]
api = "api.main:run"

[dependency-groups]
dev = ["httpx>=0.28.1,<1.0.0", "pyright", "pytest", "ruff"]

[project.urls]
Homepage = "https://github.com/luizomf/lgtm1"
Documentation = "https://github.com/luizomf/lgtm1"
Repository = "https://github.com/luizomf/lgtm1"
Issues = "https://github.com/luizomf/lgtm1/issues"

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true
show-fixes = true
indent-width = 2
exclude = ["venv", ".venv", "env", ".env", "node_modules", "__pycache__"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "T201", "INP001", "COM812"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101", "COM812"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["api"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
include = ["src", "tests"]
exclude = [
  "**/venv",
  "**/.venv",
  "**/env",
  "**/.env",
  "**/node_modules",
  "**/__pycache__",
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ['./src/api']
